backend/requirements.txt
Django==3.2
djangorestframework-jwt
djangorestframework
django-cors-headers
psycopg2
Pillow
django-markdownx
django-environ